Completely rewrote TypingGame class with new input handling
